backend/requirements.txt
# FastAPI и сервер
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# База данных
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pg8000==1.30.7

# Аутентификация и безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Валидация данных
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP клиенты
httpx==0.26.0
aiohttp==3.9.1

# Yandex Object Storage
boto3==1.34.28

# ЮKassa
yookassa==3.0.0

# Админ-панель
sqladmin==0.18.1

# Утилиты
python-dateutil==2.8.2
